+++ app/src/main/java/br/com/guiainvestimento/fragment/stock/JCPDividendFormFragment.java
@@ -95,7 +95,7 @@ public class JCPDividendFormFragment extends BaseFormFragment {
 
             // Get the stock quantity bought before the jcp/dividend is ex
             // Will be used to calculate the total R$ received of jcp/dividend
-            int stockQuantity = getStockQuantity(symbol, timestamp);
+            double stockQuantity = getStockQuantity(symbol, timestamp);
             double perStock = Double.parseDouble(mInputPerStockView.getText().toString());
             double receiveValue = stockQuantity * perStock;
             double liquidValue = receiveValue;
